+++ CHANGELOG.md
@@ -4,8 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.4.1] - 2025-03-06
 - Reinforced consistency in model dimension specification
+- Pandas column(s) selector now also accept group-by objects
+- Added pandas groupby and groupby.agg wrappers 
 
 ## [0.4.0] - 2025-02-10
 - Fixed default separator in callback after every batch of model training